Add checks before logging in
Add methods to check if the user has or not a connection. If he does not have a connection, but someone else sent him a request, the user will be redirected to the screen to accept or cancel the request, else (if he does not have a pending connection) he is redirected to the screen that suggests to connect with someone. If he has a connectio he will be redirected to the Accepted Request screen or Home screen.